requirements.txt
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
ffmpeg==1.4
fonttools==4.55.0
idna==3.4
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.23.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.28.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soundfile==0.12.1
sox==1.5.0
threadpoolctl==3.5.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.13
